Move components from reviewing to editing module
