+++ indico/modules/events/editing/client/js/components/Timeline/StateIndicator.jsx
